\name{DCODE-internal}
\alias{getConstraint_rec1}
\alias{getConstraint_rec2}
\alias{posLecturef1}
\alias{posLecturef2}
\alias{compRev}
\alias{getAllSeqs}
\alias{deleteStops}
\alias{printConstraint}
\alias{pepList}
\alias{stopCodons}
\alias{toPep}


\title{Internal DCODE Functions}
\description{
 Internal DCODE functions
}
\details{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written).
}